Fix verification tests
This commit updates the minimal in-toto root layout, removes an
unused signature, as well as updates the directory structure, putting
the links at the same level as the layout and public key. (the directory
structure should be updated in the future, but for now, getting the
tests to pass should suffice.)

Finally, it updates `docker_test.go` and `os_test.go` accordingly.

Signed-off-by: Radu M <root@radu.sh>